# docs/swagger.json
{
    "schemes": [
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "RESTful API for RaiJai Application",
        "title": "RaiJai API",
        "contact": {},
        "version": "1.0"
    },
    "host": "raijai-api.sutthiporn.dev",
    "basePath": "/api/v1",
    "paths": {
...
        "/debts": {
            "get": {
                "description": "Get all debts (optional filter by type)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "debts"
                ],
                "summary": "Get all debts",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Debt Type (LENT/BORROWED)",
                        "name": "type",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Debt"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new debt (Lent or Borrowed)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "debts"
                ],
                "summary": "Create a new debt record",
                "parameters": [
                    {
                        "description": "Debt Data",
                        "name": "debt",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Debt"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.Debt"
                        }
                    }
                }
            }
        },
        "/debts/{id}": {
            "get": {
                "description": "Get a debt by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "debts"
                ],
                "summary": "Get a debt by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Debt ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Debt"
                        }
                    }
                }
            },
            "put": {
                "description": "Update a debt by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "debts"
                ],
                "summary": "Update a debt",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Debt ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Debt Data",
                        "name": "debt",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Debt"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Debt"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete a debt by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "debts"
                ],
                "summary": "Delete a debt",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Debt ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/debts/{id}/payment": {
            "post": {
                "description": "Deduct amount from debt",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "debts"
                ],
                "summary": "Make a payment for a debt",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Debt ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Payment Amount",
                        "name": "payment",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.PaymentRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Debt"
                        }
                    }
                }
            }
        },
...
        "/upload/image": {
            "post": {
                "description": "Upload an image file (returns a mock URL)",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "system"
                ],
                "summary": "Upload an image",
                "parameters": [
                    {
                        "type": "file",
                        "description": "Image File",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Image URL",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "handlers.PaymentRequest": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "number"
                }
            }
        },
...
        "models.DateOnly": {
            "type": "object",
            "properties": {
                "time.Time": {
                    "type": "string"
                }
            }
        },
        "models.Debt": {
            "type": "object",
            "properties": {
                "autoDeduct": {
                    "type": "boolean"
                },
                "createdAt": {
                    "type": "string"
                },
                "dueDate": {
                    "$ref": "#/definitions/models.DateOnly"
                },
                "id": {
                    "type": "string"
                },
                "installmentPlan": {
                    "$ref": "#/definitions/models.InstallmentPlan"
                },
                "isInstallment": {
                    "type": "boolean"
                },
                "personName": {
                    "type": "string"
                },
                "remainingAmount": {
                    "type": "number"
                },
                "remark": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "totalAmount": {
                    "type": "number"
                },
                "type": {
                    "$ref": "#/definitions/models.DebtType"
                },
                "user": {
                    "$ref": "#/definitions/models.User"
                },
                "userId": {
                    "type": "string"
                },
                "wallet": {
                    "$ref": "#/definitions/models.Wallet"
                },
                "walletId": {
                    "type": "string"
                }
            }
        },
        "models.DebtType": {
            "type": "string",
            "enum": [
                "LENT",
                "BORROWED"
            ],
            "x-enum-varnames": [
                "DebtTypeLent",
                "DebtTypeBorrowed"
            ]
        },
...
        "models.InstallmentPlan": {
            "type": "object",
            "properties": {
                "interestRate": {
                    "type": "number"
                },
                "monthlyAmount": {
                    "type": "number"
                },
                "paidMonths": {
                    "type": "integer"
                },
                "startDate": {
                    "$ref": "#/definitions/models.DateOnly"
                },
                "totalMonths": {
                    "type": "integer"
                }
            }
        },
...
        "models.User": {
            "type": "object",
            "properties": {
                "avatarUrl": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "models.Wallet": {
            "type": "object",
            "properties": {
                "balance": {
                    "type": "number"
                },
                "color": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "currency": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "owner": {
                    "$ref": "#/definitions/models.User"
                },
                "ownerId": {
                    "type": "string"
                },
                "type": {
                    "$ref": "#/definitions/models.WalletType"
                }
            }
        },
        "models.WalletType": {
            "type": "string",
            "enum": [
                "CASH",
                "BANK",
                "CREDIT_CARD"
            ],
            "x-enum-varnames": [
                "WalletTypeCash",
                "WalletTypeBank",
                "WalletTypeCreditCard"
            ]
        }
    }
}